fix: Moved the data fetch to weather-quest-card
